requirements.txt
python-telegram-bot
sqlalchemy
python-dotenv
google-generativeai # Adicionar
dateparser
python-dateutil